refactor function of tokenization and add comment
