+++ src/analize_line.c
@@ -6,7 +6,7 @@
 /*   By: hyuncpar <hyuncpar@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/02 16:42:09 by hyuncpar          #+#    #+#             */
-/*   Updated: 2022/12/29 20:26:45 by hyuncpar         ###   ########.fr       */
+/*   Updated: 2023/01/02 16:09:54 by hyuncpar         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,6 +14,16 @@
 #include "tokenization.h"
 #include "lexer.h"
 
+/**
+ * @brief 
+ * 입력받은 line을 토큰화하고, 구문 분석을 진행한다.
+ * tokenizer: line 토큰화
+ * line_token: 공백을 기준으로 일반 문자열로 해석되는 토큰끼리 연결
+ * lexer: 구문 분석
+ * @param minishell 히어독 정보를 담기 위한 미니쉘 구조체
+ * @param input 입력받은 line
+ * @return t_token* 
+ */
 t_token	*analize_line(t_minishell *minishell, char *input)
 {
 	t_token	*tokenized;
