+++ src/tokenizer/tokenize.c
@@ -6,7 +6,7 @@
 /*   By: hyuncpar <hyuncpar@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/05 17:36:45 by hyuncpar          #+#    #+#             */
-/*   Updated: 2022/12/29 19:58:44 by hyuncpar         ###   ########.fr       */
+/*   Updated: 2023/01/02 15:55:51 by hyuncpar         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,134 +14,21 @@
 #include "tokenization.h"
 #include "testcode.h"
 
-int	tokenize_quote(t_token **token, char *input, int i, t_token_type type)
+static int	is_normal_symbol(char c)
 {
-	int		size;
-	char	c;
-
-	if (type == QUOT)
-		c = '\'';
-	else
-		c = '"';
-	size = 0;
-	while (input[i + size + 1] != c)
-		size++;
-	insert_token(token, init_token(ft_substr(input, ++i, size), type));
-	i += size + 1;
-	return (i);
+	return (c == '\'' || c == '"' || c == ' ' || c == '\\' || c == '$');
 }
 
-int	tokenize_oper(t_token **token, char *input, int i, t_token_type type)
+/**
+ * @brief 
+ * 일반 문자열로 해석되는 토큰으로 토큰화하는 함수
+ * @param token 해당 문자열 구간을 토큰화하여 정보를 담아낼 구조체
+ * @param input 첫 readline에서 입력받은 문자열
+ * @param i 	현재 토큰화가 필요한 구간의 인덱스 정보
+ * @return int 
+ */
+int	tokenize_normal(t_token **token, char *input, int i)
 {
-	int		size;
-	int		d_oper;
-	char	c ;
-
-	size = 1;
-	if (type == PIPE)
-	{
-		c = '|';
-		d_oper = DPIP;
-	}
-	else
-	{
-		c = '&';
-		d_oper = DAND;
-	}
-	if (input[i + size] == c)
-		insert_token(token, init_token(ft_substr(input, i, ++size), d_oper));
-	else
-		insert_token(token, init_token(ft_substr(input, i, size), type));
-	i += size;
-	return (i);
-}
-
-int	tokenize_redir(t_token **token, char *input, int i, int type)
-{
-	int		size;
-	int		d_oper;
-	char	c;
-
-	size = 1;
-	if (type == RIGT)
-	{
-		c = '>';
-		d_oper = DRGT;
-	}
-	else
-	{
-		c = '<';
-		d_oper = DLFT;
-	}
-	if (input[i + size] == c)
-		insert_token(token, init_token(ft_substr(input, i, ++size), d_oper));
-	else
-		insert_token(token, init_token(ft_substr(input, i, size), type));
-	i += size;
-	return (i);
-}
-
-int	tokenize_semicolon(t_token **token, char *input, int i)
-{
-	int		size;
-
-	size = 1;
-	if (input[i + size] == ';')
-		insert_token(token, init_token(ft_substr(input, i, ++size), DSEM));
-	else
-		insert_token(token, init_token(ft_substr(input, i, size), SEMC));
-	i += size;
-	return (i);
-}
-
-int	is_special_symbol(char c)
-{
-	return (c == '\'' || c == '"' || c == ' ' || c == '\\' || c == '$' \
-			|| c == '|' || c == '&' || c == '>' || c == '<' || c == '(' || c == ')');
-}
-
-int	tokenize_etc(t_token **token, char *input, int i, int type)
-{
-	int		size;
-
-	size = 1;
-	if (type == SPCE)
-		while (input[i + size] == ' ')
-			size++;
-	else if (type == BSLH)
-		i++;
-	else if (type == DOLR)
-	{
-		i++;
-		while (input[i + size] && !is_special_symbol(input[i + size]))
-			size++;
-		if (size == 1)
-		{
-			insert_token(token, init_token(ft_strdup("$"), NORM));
-			return (++i);
-		}
-		size--;
-	}
-	insert_token(token, init_token(ft_substr(input, i, size), type));
-	i += size;
-	return (i);
-}
-
-int	tokenize_prnth_or_newline(t_token **token, char *input, int i)
-{
-	if (input[i] == ')')
-		insert_token(token, init_token(ft_strdup(")"), PARENTHESIS_RIGHT));
-	else if (input[i] == '(')
-		insert_token(token, init_token(ft_strdup("("), PARENTHESIS_LEFT));
-	else
-		insert_token(token, init_token(ft_strdup("\n"), NEW_LINE));
-	return (++i);
-}
-
-int	tokenize_line(t_token **token, char *input, int index, int i)
-{
-	if (i - index)
-		insert_token(token, init_token(ft_substr(input, index, i - index), NORM));
 	if (input[i] == '\'')
 		i = tokenize_quote(token, input, i, QUOT);
 	else if (input[i] == '"')
@@ -152,7 +39,20 @@ int	tokenize_line(t_token **token, char *input, int index, int i)
 		i = tokenize_etc(token, input, i, BSLH);
 	else if (input[i] == '$')
 		i = tokenize_etc(token, input, i, DOLR);
-	else if (input[i] == '|')
+	return (i);
+}
+
+/**
+ * @brief 
+ * 특별한 연산이 필요한 토큰으로 토근화하는 함수
+ * @param token 해당 문자열 구간을 토큰화하여 정보를 담아낼 구조체
+ * @param input 첫 readline에서 입력받은 문자열
+ * @param i 	현재 토큰화가 필요한 구간의 인덱스 정보
+ * @return int 
+ */
+int	tokenize_special(t_token **token, char *input, int i)
+{
+	if (input[i] == '|')
 		i = tokenize_oper(token, input, i, PIPE);
 	else if (input[i] == '&')
 		i = tokenize_oper(token, input, i, NORM);
@@ -164,5 +64,28 @@ int	tokenize_line(t_token **token, char *input, int index, int i)
 		i = tokenize_prnth_or_newline(token, input, i);
 	else if (input[i] == ';')
 		i = tokenize_semicolon(token, input, i);
+	return (i);
+}
+
+/**
+ * @brief 
+ * 토큰화가 필요한 문자가 발견되면, 그 전까지 토큰화 되지 않은 문자열은 normal 타입으로 토큰화한다.
+ * 그 이후, 일반 문자열로 해석되는 토큰인지, 특별한 연산이 필요한 토큰인지 확인한다.
+ * @param token 해당 문자열 구간을 토큰화하여 정보를 담아낼 구조체
+ * @param input 첫 readline에서 입력받은 문자열
+ * @param index 현재까지 토큰화가 완료된 구간의 인덱스 정보
+ * @param i 	현재 토큰화가 필요한 구간의 인덱스 정보
+ * @return int 
+ */
+
+int	tokenize_line(t_token **token, char *input, int index, int i)
+{
+	if (i - index)
+		insert_token(token, \
+		init_token(ft_substr(input, index, i - index), NORM));
+	if (is_normal_symbol(input[i]))
+		i = tokenize_normal(token, input, i);
+	else
+		i = tokenize_special(token, input, i);
 	return (i);
 }
