+++ src/tokenizer/tokenizer.c
@@ -6,7 +6,7 @@
 /*   By: hyuncpar <hyuncpar@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/02 16:42:09 by hyuncpar          #+#    #+#             */
-/*   Updated: 2022/12/28 15:49:42 by hyuncpar         ###   ########.fr       */
+/*   Updated: 2023/01/02 15:49:18 by hyuncpar         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,6 +14,12 @@
 #include "tokenization.h"
 #include "testcode.h"
 
+/**
+ * @brief 
+ * 토큰화가 필요한 문자인지 검사한다.
+ * @param c 
+ * @return int 
+ */
 static int	is_valid_char(char c)
 {
 	return (c == '\'' || c == '"' || c == ' ' || c == '\\' || c == '$' \
@@ -21,6 +27,15 @@ static int	is_valid_char(char c)
 			|| c == '(' || c == ')' || c == '\n' || c == ';');
 }
 
+/**
+ * @brief 
+ * 입력받은 line을 토큰화하여 연결 리스트에 담아낸다.
+ * 한 글자씩 검사하여 토큰화가 필요한 문자가 발견되면
+ * 현재까지 검사한 인덱스 정보와 토큰을 담아낼 구조체를
+ * tokenize_line 함수에 인자로 전달한다.
+ * @param input 
+ * @return t_token* 
+ */
 t_token	*tokenizer(char *input)
 {
 	int		end_index;
@@ -39,11 +54,6 @@ t_token	*tokenizer(char *input)
 		}
 		else
 			end_index++;
-		if (end_index == TOKEN_ERROR)	//TODO minishell status change
-		{
-			free_tokens(token);
-			return (token);
-		}
 	}
 	return (token);
 }
